Stylistic changes
