+++ Assets/Libraries/DiscordGameSDK.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 74f135a4d5fd254419d86c905840e656
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
