+++ Assets/Script/DiscordController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 523c0f56c7bfffa45adfc8894b578416
+guid: cb438f304d0d4064fb1de229c53becde
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
